vim-patch:8.1.1734: the evalfunc.c file is too big
Problem:    The evalfunc.c file is too big.
Solution:   Move some functions to other files.

https://github.com/vim/vim/commit/29b7d7a9aac591f920edb89241c8cde27378e50b
